requirements.txt
Django==4.1
Pillow==10.0.1
mysqlclient==2.2.0